Worked on interface

+++ src/main/java/net/berndreiss/zentodo/data/User.java
@@ -25,7 +25,7 @@ public class User {
     private boolean enabled = false;
 
     @Column (nullable = false)
-    private long device;
+    private int device;
 
     @Column (nullable = false)
     private int profile;
@@ -56,11 +56,11 @@ public class User {
         this.userName = userName;
     }
 
-    public User(String email, long device){
+    public User(String email, int device){
         this.email = email;
         this.device = device;
     }
-    public User(String email, String username, long device){
+    public User(String email, String username, int device){
         this.email = email;
         this.userName = username;
         this.device = device;
@@ -98,9 +98,9 @@ public class User {
         this.enabled = enabled;
     }
 
-    public long getDevice(){return device;}
+    public int getDevice(){return device;}
 
-    public void setDevice(long device){this.device = device;}
+    public void setDevice(int device){this.device = device;}
 
     public int getProfile(){return profile;}
